+++ genieclust/inequity.pyx
@@ -12,7 +12,7 @@ Inequity (inequality) measures
 
 # ############################################################################ #
 #                                                                              #
-#   Copyleft (C) 2020-2021, Marek Gagolewski <https://www.gagolewski.com>      #
+#   Copyleft (C) 2020-2022, Marek Gagolewski <https://www.gagolewski.com>      #
 #                                                                              #
 #                                                                              #
 #   This program is free software: you can redistribute it and/or modify       #
@@ -74,6 +74,7 @@ cpdef double gini_index(np.ndarray[T] x, bint is_sorted=False):
     --------
 
     genieclust.inequity.bonferroni_index : The normalised Bonferroni index
+    genieclust.inequity.devergottini_index : The normalised De Vergottini index
 
 
     Notes
@@ -101,13 +102,12 @@ cpdef double gini_index(np.ndarray[T] x, bint is_sorted=False):
 
     where :math:`\\sigma` is an ordering permutation of :math:`(x_1,\\dots,x_n)`.
 
-
-    Both the Gini and Bonferroni indices can be used to quantify the "inequity"
-    of a numeric sample. They can be perceived as measures of data dispersion.
-    For constant vectors (perfect equity), the indices yield values of 0.
-    Vectors with all elements but one equal to 0 (perfect inequity),
-    are assigned scores of 1.
-    Both indices follow the Pigou-Dalton principle (are Schur-convex):
+    The Gini, Bonferroni, and De Vergottini indices can be used to quantify
+    the "inequity" of a numeric sample. They can be perceived as measures
+    of data dispersion. For constant vectors (perfect equity), the indices
+    yield values of 0. Vectors with all elements but one equal to 0
+    (perfect inequity), are assigned scores of 1.
+    They follow the Pigou-Dalton principle (are Schur-convex):
     setting :math:`x_i = x_i - h` and :math:`x_j = x_j + h` with :math:`h > 0`
     and :math:`x_i - h \\geq  x_j + h` (taking from the "rich" and giving away
     to the "poor") decreases the inequity.
@@ -183,6 +183,7 @@ cpdef double bonferroni_index(np.ndarray[T] x, bint is_sorted=False):
     See Also
     --------
 
+    genieclust.inequity.devergottini_index : The normalised De Vergottini index
     genieclust.inequity.gini_index : The normalised Gini index
 
 
@@ -240,3 +241,83 @@ cpdef double bonferroni_index(np.ndarray[T] x, bint is_sorted=False):
     else: x = np.array(x, dtype=x.dtype, copy=False, order="C") # assure c_contiguity
 
     return c_inequity.Cbonferroni_sorted(&x[0], x.shape[0])
+
+
+
+cpdef double devergottini_index(np.ndarray[T] x, bint is_sorted=False):
+    """
+    genieclust.inequity.devergottini_index(x, is_sorted=False)
+
+    Computes the normalised De Vergottini index
+
+
+    Parameters
+    ----------
+
+    x : ndarray
+        A vector with non-negative elements.
+
+    is_sorted : bool
+        Indicates if `x` is already sorted increasingly.
+
+
+    Returns
+    -------
+
+    double
+        The value of the inequity index, a number in [0,1].
+
+
+    See Also
+    --------
+
+    genieclust.inequity.bonferroni_index : The normalised Bonferroni index
+    genieclust.inequity.gini_index : The normalised Gini index
+
+
+    Notes
+    -----
+
+    The normalised De Vergottini index is given by:
+
+    .. math::
+
+        \\frac{1}{\\sum_{i=2}^n \\frac{1}{i}}
+        \\left(
+        \\frac{ \\sum_{i=1}^n \\left(
+            \\sum_{j=i}^{n} \\frac{1}{j}
+        \\right) x_{\\sigma(n-i+1)} }{\\sum_{i=1}^{n} x_i} - 1
+        \\right)
+
+    where :math:`\\sigma` is an ordering permutation of :math:`(x_1,\\dots,x_n)`.
+
+    Time complexity is :math:`O(n)` for sorted data.
+
+
+    Examples
+    --------
+
+    No inequality (perfect equality):
+
+    >>> round(genieclust.inequity.devergottini_index(np.r_[2, 2,  2, 2, 2]), 2)
+    0.0
+
+    One has it all (total inequity):
+
+    >>> round(genieclust.inequity.devergottini_index(np.r_[0, 0, 10, 0, 0]), 2)
+    1.0
+
+    Give to the poor, take away from the rich:
+
+    >>> round(genieclust.inequity.devergottini_index(np.r_[7, 0,  3, 0, 0]), 2)
+    0.77
+
+    Robinhood even more:
+
+    >>> round(genieclust.inequity.devergottini_index(np.r_[6, 0,  3, 1, 0]), 2)
+    0.65
+    """
+
+    if not is_sorted: x = np.sort(x)
+    else: x = np.array(x, dtype=x.dtype, copy=False, order="C") # assure c_contiguity
+    return c_inequity.Cdevergottini_sorted(&x[0], x.shape[0])